% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job.R
\name{create_job}
\alias{create_job}
\title{Create a bash file to submit a script to SLURM}
\usage{
create_job(outfile, jobname, script_lines, opts = list())
}
\arguments{
\item{outfile}{[string] The path / filename to save the bash file.}

\item{jobname}{[string] What should the jobname be on slurm?}

\item{script_lines}{[string] A command that should be executed on
the terminal on the cluster machine.}

\item{opts}{[list] Batch submission options. See merge_defaults for
default arguments.}
}
\description{
Create a bash file to submit a script to SLURM
}

