% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job.R
\name{merge_defaults}
\alias{merge_defaults}
\title{Merge default batch options}
\usage{
merge_defaults(opts)
}
\arguments{
\item{opts}{[list] A list partially specifying options expected by
create_job.}
}
\value{
opts [list] The original opts list with unspecified opts
  filled in with defaults.
}
\description{
This will merge in defaults to the opts in create_job, so that we
don't need to specify all the options every time.
}

