% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job.R
\name{output_path}
\alias{output_path}
\title{Function to concatenate path to file}
\usage{
output_path(dir)
}
\arguments{
\item{dir}{[string] The parent directory the returned function
should concatenate names to.}
}
\value{
f [function] A function that, when called on a filename,
  concatenates dir over that filename.
}
\description{
This is a wrapper function that simplifies specifying long paths to
files. It returns a function that, when called on a filename, returns
the path to the original directory + that filename.
}

